+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- New optional ``label_keys`` parameter for ``histogram()`` metrics
+
 ### Changed
 
 ### Fixed